requirements.txt
python-telegram-bot==20.0
apscheduler==3.10.0
python-dotenv==1.0.0
